+++ src/scripts/infuse/infuse.module.js
@@ -1,0 +1,8 @@
+import angular from 'angular';
+
+import { InfuseComponent } from './infuse.component';
+
+export default angular
+  .module('infuseModule', [])
+  .component('infuse', InfuseComponent)
+  .name;