+++ src/scripts/move-popup/dimMovePopup.directive.js
@@ -1,6 +1,5 @@
 import angular from 'angular';
 import template from './dimMovePopup.directive.html';
-import infuseTemplate from 'app/views/infuse.html';
 
 angular.module('dimApp')
   .directive('dimMovePopup', MovePopup);
@@ -74,11 +73,13 @@ function MovePopupController($scope, dimStoreService, ngDialog, $timeout, dimSet
 
     // Open the infuse window
     ngDialog.open({
-      template: infuseTemplate,
-      plain: true,
+      template: '<infuse source="item"></infuse>',
       className: 'app-settings',
       appendClassName: 'modal-dialog',
-      data: item
+      controller: function($scope) {
+        'ngInject';
+        $scope.item = item;
+      }
     });
   };